lockn-score/requirements.txt
ultralytics>=8.1.0
opencv-python>=4.9.0
numpy>=1.26.0
sounddevice>=0.4.6
requests>=2.32.0
websocket-client>=1.7.0

# Audio (PANNs)
panns-inference>=0.1.1
sounddevice>=0.4.7
torch>=2.2.0
torchaudio>=2.2.0
fastapi
pydantic
uvicorn
websockets
